server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
pdfplumber==0.10.3
python-docx==1.1.0
aiofiles==23.2.1
